backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "EmanHamdyMohamed",email = "emanhamdy645@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.104.0,<0.105.0)",
    "uvicorn (>=0.23.2,<0.24.0)",
    "motor (>=3.3.1,<4.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "pymongo (>=4.12.0,<5.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aiosmtplib (>=4.0.1,<5.0.0)",
    "twilio (>=9.5.2,<10.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "python-telegram-bot (>=22.0,<23.0)"
]

# Add this section to specify your package
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
